+++ x/btcstaking/keeper/genesis_test.go
@@ -103,10 +103,10 @@ func TestExportGenesis(t *testing.T) {
 				DelBtcPk: del.BtcPk,
 			}
 
-			// record event that the BTC delegation will become unbonded at EndHeight-w
+			// record event that the BTC delegation will become expired (unbonded) at EndHeight-w
 			unbondedEvent := types.NewEventPowerDistUpdateWithBTCDel(&types.EventBTCDelegationStateUpdate{
 				StakingTxHash: stakingTxHash.String(),
-				NewState:      types.BTCDelegationStatus_UNBONDED,
+				NewState:      types.BTCDelegationStatus_EXPIRED,
 			})
 
 			// events